% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIAs.R
\name{FIAs}
\alias{FIAs}
\title{Feature Importance Algorithms with Incremental Feature Selection}
\usage{
FIAs(
  data,
  algorithms = c("mrmr", "PVE", "Cor", "RF", "ridge", "gwas"),
  n_features = NULL,
  n_folds = 10,
  output_path = "./",
  seed = 1,
  mrmr.threads = NULL,
  ridge.lambda = 1:300,
  ridge.metrics = "regr.rsq",
  ridge.evals = NULL,
  gwas.myY,
  gwas.myGM,
  gwas.nPCA = 3,
  gwas.model = c("FarmCPU"),
  metrics = c("regr.rsq", "regr.mse", "time_both", "regr.rmse", "regr.bias", "regr.ktau",
    "regr.mae", "regr.mape", "regr.maxae", "regr.medae", "regr.medse", "regr.msle",
    "regr.pbias", "regr.rae", "regr.rmsle", "regr.rrse", "regr.rse", "regr.sae",
    "regr.smape", "regr.srho", "regr.sse")
)
}
\arguments{
\item{data}{Data frame format, the first column is the target variable (trait value), and the remaining columns are SNP features. The encoding format of these features can be 0, 1, 2 or -1, 0, 1.}

\item{algorithms}{Character vector format, default is c("mrmr", "PVE", "Cor","RF","ridge","gwas"), case insensitive, these algorithms perform feature selection and input the obtained features into RF to return the evaluation index. In particular, when gwas is selected, it is necessary to read the two data frame parameters gwas.myY and gwas.myGM related to the algorithm.}

\item{n_features}{Numeric format, with the default set to the total number of features. This parameter defines the maximum number of features to be selected, effectively determining the value of K in the Top K features. The features are ranked by importance and sequentially added to the model one by one until the number of selected features reaches the specified n_features (i.e., incremental feature selection). This process will use the maximum rsq of the model performance in 1:n_features as the criterion for feature selection.}

\item{n_folds}{Numeric format, default is 10, specifies the cross validation fold of the random forest model.}

\item{output_path}{Output result file, default is the path where getwd() is located.}

\item{seed}{Numeric format, random seed for random forest model training.}

\item{mrmr.threads}{Numeric format, the number of threads of the mRMR algorithm, the default is the maximum number of threads on the computer minus one.}

\item{ridge.lambda}{A continuous integer vector specifying the search range for lambda in the ridge algorithm. The default value is 1:300.}

\item{ridge.metrics}{Character format, performance evaluation index of ridge algorithm, default is "regr.rsq".}

\item{ridge.evals}{Numeric format, the maximum number of evaluations during hyperparameter tuning, that is, the maximum number of lambda values the tuner will test. If NULL, the number of evaluations defaults to the range (max - min) of \code{ridge.lambda}.}

\item{gwas.myY}{A data frame required when using the GWAS algorithm. The first column should contain genotype names, and the second column should contain the trait values.
The genotype names in the first column must match the genotype names in the feature column of \code{data}.}

\item{gwas.myGM}{A data frame required when using the GWAS algorithm. It must have three columns: the first for SNP identifiers, the second for chromosome numbers, and the third for the corresponding physical positions.}

\item{gwas.nPCA}{Numerical format, the number of principal components (PCA) included in the GWAS model, as used in the GAPIT package.}

\item{gwas.model}{Character format, one or more GWAS models supported by the GAPIT package, such as "FarmCPU", "MLM", "GLM", etc. Default is "FarmCPU".}

\item{metrics}{Character vector specifying the performance metrics used to evaluate the random forest model.
The final variable selection is based on "regr.rsq".
Defaults to 21 common regression metrics, including "regr.rsq", "regr.mse", "regr.rmse", among others.}
}
\value{
A list with three components:
\describe{
\item{data}{The input data frame.}
\item{parameters}{A list of input parameters and settings.}
\item{results}{A list of results for each specified feature importance algorithm, including the ranked features, model object, runtime, memory usage, and cross-validation outcomes.}
}
}
\description{
A comprehensive SNP screening approach was implemented by integrating multiple feature importance measures, including mRMR, SNP's PVE, Pearson correlation between SNPs and traits, permutation-based random forest importance, absolute coefficients from ridge regression, and GWAS p-values, in combination with incremental feature selection.
}
\details{
This function performs feature selection by combining multiple feature importance algorithms and incremental feature selection (IFS).
First, features are ranked by importance using the specified algorithms.
Next, features are incrementally added into the predictive model (random forest by default) in order of their importance.
The model performance (evaluated via rsq) is tracked across 1:n_features, and the feature subset achieving the maximum rsq is selected as the final set of informative SNPs.
}
\examples{
\dontrun{
# Running the Feature Importance Algorithms with Incremental Feature Selection
FIAs_result <- FIAs(
  data = WW1768,
  gwas.myY = gwas.myY,
  gwas.myGM = gwas.myGM,
  n_folds = 10,
  n_features = 500,
  ridge.lambda = 1:300,
  algorithms = c("mrmr", "PVE", "Cor", "RF", "ridge", "gwas")
)
rank_fias(data = FIAs_result)
}

}
\references{
Bermingham ML, Pong-Wong R, Spiliopoulou A, et al (2015) Application of high-dimensional feature selection: evaluation for genomic prediction in man. Scientific reports 5(1):10312.
\url{https://doi.org/10.1038/srep10312}

Wang J, Zhang Z (2021) GAPIT version 3: boosting power and accuracy for genomic association and prediction. Genomics, proteomics & bioinformatics 19(4):629-640.
\url{https://doi.org/10.1016/j.gpb.2021.08.005}

Lang M, Binder M, Richter J, et al (2019) mlr3: A modern object-oriented machine learning framework in R. Journal of Open Source Software 4(44):1903.
\url{https://doi.org/10.21105/joss.01903}

Peng H, Long F, Ding C (2005) Feature selection based on mutual information criteria of max-dependency, max-relevance, and min-redundancy. IEEE Transactions on pattern analysis and machine intelligence 27(8):1226-1238.
\url{https://doi.org/10.1109/TPAMI.2005.159}

Breiman L (2001) Random forests. Machine learning 45(1):5-32.
\url{https://doi.org/10.1023/A:1010933404324}

Hoerl AE, Kennard RW (1970) Ridge regression: Biased estimation for nonorthogonal problems. Technometrics 12(1):55-67.
\url{https://doi.org/10.1080/00401706.1970.10488634}

Friedman JH, Hastie T, Tibshirani R (2010) Regularization paths for generalized linear models via coordinate descent. Journal of statistical software, 33(1):1-22.
\url{https://doi.org/10.18637/jss.v033.i01}

Wright MN, Wager S, Probst P (2020) Ranger: A fast implementation of random forests. R package version 0.12 1:730.
\url{https://doi.org/10.18637/jss.v077.i01}
}
\seealso{
\code{\link{FIAs_W},\link{rank_fias}}
}
