% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIAs_W.R
\name{FIAs_W}
\alias{FIAs_W}
\title{FIAs weighting algorithm}
\usage{
FIAs_W(data, top_n = NULL, weight_step = 0.5, output_file = "FIAs_W.xlsx")
}
\arguments{
\item{data}{The result list generated by FIAs() is used as the input of FIAs_W().}

\item{top_n}{Numeric format, its value is between 1:n_features, where n_features is the corresponding parameter of FIAs(). FIAs_W() will generate a new feature ranking based on the fluctuating weights, and top_n determines that FIAs_W() will dynamically select features between 1:top_n and compare the rsq indicator to find the corresponding weight.}

\item{weight_step}{Numeric format, determines the step size when traversing weights.}

\item{output_file}{File name ending with .xlsx, output weight results, optimal weight, cross-validation performance, feature ranking, time, and memory usage.}
}
\value{
A list with three components:
\describe{
\item{\code{data}}{The input data frame.}
\item{\code{parameters}}{A list of input parameters and settings.}
\item{\code{results}}{A list of results from each specified feature importance algorithm, including ranked features and model objects.}
}
Compared with the result obtained by FIAs(), the FIAs_W algorithm is added to the result.
}
\description{
Weighting the top three FIAs to achieve performance improvement.
}
\details{
This function enhances feature selection by weighting the top three feature importance algorithms (FIAs) obtained from FIAs(). It dynamically combines their rankings using varying weights within a specified range, evaluates model performance (R²) via cross-validation for each weight combination on the top features up to \code{top_n}, and selects the optimal weights that maximize predictive accuracy. Subsequently, incremental feature selection (IFS) is conducted on the weighted ranking to determine the final informative feature subset. The output includes optimal weights, performance metrics, feature rankings, runtime, and memory usage.
}
\examples{
\dontrun{
FIAs_result <- FIAs(
  data = WW1768,
  gwas.myY = gwas.myY,
  gwas.myGM = gwas.myGM,
  n_folds = 10,
  n_features = 500,
  ridge.lambda = 1:300,
  algorithms = c("mrmr", "PVE", "Cor", "RF", "ridge", "gwas")
)
FIAs_W_result <- FIAs_W(data = FIAs_result,weight_step = 0.1)
rank_fias(FIAs_W_result)
}

}
\seealso{
\code{\link{FIAs},\link{rank_fias}}
}
