% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwo_plot.R
\name{gwo_plot}
\alias{gwo_plot}
\title{Plot the performance evolution of GWO/PGWO feature selection}
\usage{
gwo_plot(data)
}
\arguments{
\item{data}{A list returned by \code{gwo_fs()}, containing at least the element
\code{performance}, a data frame with columns \code{iteration} and \code{regr.rsq}.}
}
\value{
A ggplot2 object showing three lines—mean, median, and max R²—across GWO/PGWO iterations.
}
\description{
\code{gwo_plot()} visualizes the performance evolution over iterations for results
generated by \code{\link{gwo_fs}}.It shows the trends of mean R², median R²,
and maximum R² across generations,allowing users to evaluate the convergence behavior
of the Grey Wolf Optimizer (GWO)  or Priority-Guided Grey Wolf Optimizer (PGWO) in the feature selection process.
}
\examples{
\dontrun{
union <- feature_union(FIAs_W_result$results$FIAs_W,abess_result,Boruta_result,PL_result,EN_result,mmpc_result,data = WW1768)
feature_union <- union$union_features
GWO_result <- gwo_fs(
  data = WW1768,
  feature_cols = feature_union,
  seed = 1,
  control = list(numPopulation = 10, maxIter = 20, elite = 1),
  output_file = "GWO.csv"
)

gwo_plot(GWO_result)
}

}
\seealso{
\code{\link{gwo_fs}}
}
