% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmpc_regression.R
\name{mmpc_regression}
\alias{mmpc_regression}
\title{Feature Selection Using MMPC regression}
\usage{
mmpc_regression(
  data,
  seed = 1,
  n_folds,
  max_k_range = 1:10,
  threshold = 0.05,
  ncores = NULL,
  metrics = c("regr.rsq", "regr.mse", "time_both", "regr.rmse", "regr.bias", "regr.ktau",
    "regr.mae", "regr.mape", "regr.maxae", "regr.medae", "regr.medse", "regr.msle",
    "regr.pbias", "regr.rae", "regr.rmsle", "regr.rrse", "regr.rse", "regr.sae",
    "regr.smape", "regr.srho", "regr.sse"),
  output_file = "MMPC_result.xlsx"
)
}
\arguments{
\item{data}{Data frame format, the first column is the target variable (trait value), and the remaining columns are SNP features. The encoding format of these features can be 0, 1, 2 or -1, 0, 1.}

\item{seed}{Numeric format, random seed for random forest model training.}

\item{n_folds}{Numeric format, default is 10, specifies the cross validation fold of the random forest model.}

\item{max_k_range}{Integer vector. A sequence of values for the \code{max_k} parameter in the MMPC algorithm. Each value defines the maximum size of the conditioning set used in conditional independence tests. Increasing \code{max_k} allows more complex dependency structures to be detected but increases computation time.}

\item{threshold}{Numeric. The significance level (usually between 0 and 1, e.g., 0.05) used in the MMPC algorithm to determine whether a variable should be included. A lower value makes the selection more stringent.}

\item{ncores}{Integer. Number of processor cores to use for parallel computation during feature selection. If \code{NULL}, it defaults to one less than the number of available physical cores.}

\item{metrics}{Character vector specifying the performance metrics used to evaluate the random forest model.
The final variable selection is based on "regr.rsq".
Defaults to 21 common regression metrics, including "regr.rsq", "regr.mse", "regr.rmse", among others.}

\item{output_file}{Character string, default "MMPC_result.xlsx". Name of the Excel file to save the MMPC feature statistics and model evaluation results.}
}
\value{
A list containing:
\describe{
\item{scores}{A data frame of cross-validation scores for each support size and fold.}
\item{aggregates}{A data frame of aggregated scores for each support size.}
\item{feature_selection}{A character vector of confirmed important features selected by MMPC.}
\item{best_max_k}{The \code{max_k} value yielding the best R-squared.}
\item{rsq}{The best cross-validated R-squared value.}
\item{time_mins}{A numeric value indicating the total runtime of the function in minutes.}
\item{memory_bytes}{A numeric value representing the approximate memory usage (in bytes) during execution, computed via \code{pryr::mem_used()}.}
}
}
\description{
Applies the MMPC (Max-Min Parents and Children) algorithm for feature selection in a regression context.
For each specified value of \code{max_k}, it selects features, builds a regression model using \code{ranger},
and evaluates performance through cross-validation
}
\details{
The MMPC (Max-Min Parents and Children) algorithm is a constraint-based feature selection method
that identifies relevant variables by iteratively testing conditional independencies,
aiming to find the direct causes and effects (parents and children) of a target variable.
The parameter \code{max_k} controls the maximum size of the conditioning set in these tests,
allowing the detection of increasingly complex dependency structures as \code{max_k} increases,
at the cost of greater computational time.

This function applies MMPC for feature selection across a range of \code{max_k} values.
For each \code{max_k}, it selects relevant features from the input dataset,
then fits a random forest regression model using \code{ranger} on the selected features.
The model's predictive performance is evaluated by k-fold cross-validation according to multiple metrics.
}
\examples{
\dontrun{
mmpc_result <- mmpc_regression(data = WW1768, n_folds = 10,max_k_range = 1:10,threshold = 0.05,seed = 1)
}

}
\references{
Tsamardinos I, Brown LE, Aliferis CF (2006) The max-min hill-climbing Bayesian network structure learning algorithm. Machine learning 65(1):31-78.
\url{https://doi.org/10.1007/s10994-006-6889-7}
}
\seealso{
\code{\link{boruta_regression},\link{Cluster_priority_lasso},\link{auto_abess},\link{auto_abess}}}
}
