% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_fias.R
\name{rank_fias}
\alias{rank_fias}
\title{Ranking Feature Selection Algorithms and Their Feature Rankings from FIAs or FIAs_W Results}
\usage{
rank_fias(data)
}
\arguments{
\item{data}{The result list generated by \code{FIAs()} or \code{FIAs_W()} is used as the input of \code{rank_fias()}.}
}
\value{
A list with three components:
\describe{
\item{\code{Performance_ranking}}{A data frame summarizing the prediction performance of each feature selection algorithm, including the maximum R² value and the number of features used.}
\item{\code{feature_rank_matrix}}{A data frame showing the feature ranking results for each algorithm. Each column corresponds to an algorithm, and each row gives the rank of a specific feature.}
\item{\code{metadata}}{A list containing metadata information, including the generation timestamp, the total number of features ranked, and the list of algorithms used.}
}
}
\description{
Rank the results of FIAs according to "regr.rsq" and show the top three algorithms and feature rankings.
}
\details{
This function takes the output list from \code{FIAs()} or \code{FIAs_W()}, which contains
the results of multiple feature selection algorithms evaluated on regression tasks.
It ranks these algorithms based on their maximum cross-validated R² values ("regr.rsq"),
then extracts and summarizes the top three algorithms in terms of prediction performance.

Specifically, it:
\itemize{
\item Verifies the input data structure and necessary components.
\item Extracts the maximum R² and corresponding number of features used by each algorithm.
\item Sorts algorithms descendingly by their maximum R² value to determine performance ranking.
\item Constructs a feature ranking matrix for the top three algorithms, where each column
represents an algorithm and each row represents a feature's rank within that algorithm.
}

The output provides a comprehensive overview of algorithm performance and feature importance rankings,
facilitating comparison and interpretation of feature selection results across multiple methods.
}
\examples{
\dontrun{
FIAs_result <- FIAs(
  data = WW1768,
  gwas.myY = gwas.myY,
  gwas.myGM = gwas.myGM,
  n_folds = 10,
  n_features = 500,
  ridge.lambda = 1:300,
  algorithms = c("mrmr", "PVE", "Cor", "RF", "ridge", "gwas")
)
rank_fias(data = FIAs_result)

FIAs_W_result <- FIAs_W(data = FIAs_result,weight_step = 0.1)
rank_fias(data = FIAs_W_result)
}

}
\seealso{
\code{\link{FIAs},\link{FIAs_W}}
}
